# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "infinitas hosted registry",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/library/{object_id}/releases": {
      "get": {
        "tags": [
          "library"
        ],
        "summary": "Library Releases",
        "description": "List releases for a library object with pagination.\n\nPath Parameters:\n    object_id: The library object ID\n\nQuery Parameters:\n    skip: Number of items to skip (default: 0, max: 10000)\n    limit: Maximum items to return (default: 20, max: 100)",
        "operationId": "library_releases_api_v1_library__object_id__releases_get",
        "parameters": [
          {
            "name": "object_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Object Id"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 10000,
              "minimum": 0,
              "description": "Number of items to skip",
              "default": 0,
              "title": "Skip"
            },
            "description": "Number of items to skip"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Maximum number of items to return",
              "default": 20,
              "title": "Limit"
            },
            "description": "Maximum number of items to return"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "infinitas_auth_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Infinitas Auth Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryReleaseListView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "SessionCookie": []
          }
        ]
      }
    },
...
    "/api/v1/versions/{version_id}/releases": {
      "post": {
        "tags": [
          "release"
        ],
        "summary": "Create Release",
        "operationId": "create_release_api_v1_versions__version_id__releases_post",
        "parameters": [
          {
            "name": "version_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Version Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "infinitas_auth_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Infinitas Auth Token"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "SessionCookie": []
          }
        ]
      }
    },
    "/api/v1/releases/{release_id}": {
      "get": {
        "tags": [
          "release"
        ],
        "summary": "Get Release",
        "operationId": "get_release_api_v1_releases__release_id__get",
        "parameters": [
          {
            "name": "release_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Release Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "infinitas_auth_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Infinitas Auth Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "SessionCookie": []
          }
        ]
      }
    },
    "/api/v1/releases/{release_id}/artifacts": {
      "get": {
        "tags": [
          "release"
        ],
        "summary": "List Release Artifacts",
        "operationId": "list_release_artifacts_api_v1_releases__release_id__artifacts_get",
        "parameters": [
          {
            "name": "release_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Release Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "infinitas_auth_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Infinitas Auth Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactListView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "SessionCookie": []
          }
        ]
      }
    },
...
    "/library/{object_id}/releases/{release_id}": {
      "get": {
        "summary": "Library Release Page",
        "operationId": "library_release_page_library__object_id__releases__release_id__get",
        "parameters": [
          {
            "name": "object_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Object Id"
            }
          },
          {
            "name": "release_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Release Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ArtifactListView": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ArtifactView"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "default": 0
          }
        },
        "type": "object",
        "title": "ArtifactListView"
      },
      "ArtifactView": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "release_id": {
            "type": "integer",
            "title": "Release Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "storage_uri": {
            "type": "string",
            "title": "Storage Uri"
          },
          "sha256": {
            "type": "string",
            "title": "Sha256"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "release_id",
          "kind",
          "storage_uri",
          "sha256",
          "size_bytes",
          "created_at"
        ],
        "title": "ArtifactView"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LibraryReleaseListView": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/LibraryReleaseView"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "default": 0
          },
          "skip": {
            "type": "integer",
            "title": "Skip",
            "default": 0
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "default": 20
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More",
            "default": false
          }
        },
        "type": "object",
        "title": "LibraryReleaseListView"
      },
      "LibraryReleaseView": {
        "properties": {
          "release_id": {
            "type": "integer",
            "title": "Release Id"
          },
          "version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "ready_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ready At"
          },
          "visibility": {
            "$ref": "#/components/schemas/LibraryVisibilityView"
          },
          "content_digest": {
            "type": "string",
            "title": "Content Digest"
          },
          "metadata_digest": {
            "type": "string",
            "title": "Metadata Digest"
          }
        },
        "type": "object",
        "required": [
          "release_id",
          "state",
          "created_at",
          "visibility",
          "content_digest",
          "metadata_digest"
        ],
        "title": "LibraryReleaseView"
      },
      "LibraryVisibilityView": {
        "properties": {
          "audience_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audience Type"
          },
          "listing_mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Listing Mode"
          },
          "install_mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Install Mode"
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          }
        },
        "type": "object",
        "title": "LibraryVisibilityView"
      },
...
      "ReleaseView": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "skill_version_id": {
            "type": "integer",
            "title": "Skill Version Id"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "format_version": {
            "type": "string",
            "title": "Format Version"
          },
          "manifest_artifact_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Manifest Artifact Id"
          },
          "bundle_artifact_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bundle Artifact Id"
          },
          "signature_artifact_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signature Artifact Id"
          },
          "provenance_artifact_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provenance Artifact Id"
          },
          "created_by_principal_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By Principal Id"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "ready_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ready At"
          },
          "platform_compatibility": {
            "additionalProperties": true,
            "type": "object",
            "title": "Platform Compatibility",
            "default": {}
          }
        },
        "type": "object",
        "required": [
          "id",
          "skill_version_id",
          "state",
          "format_version",
          "created_at"
        ],
        "title": "ReleaseView"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      },
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "infinitas_auth_token"
      }
    }
  }
}